ghc-tag-core: use GhcTags prefix for all modules
