+++ ghc-tags-plugin/src/Plugin/GhcTags.hs
@@ -49,14 +49,14 @@ import           HsSyn (HsModule (..))
 import qualified Outputable as Out
 import qualified PprColour
 
+import           GhcTags.Generate
+import           GhcTags.Tag
+import           GhcTags.Stream
+import qualified GhcTags.CTags as CTags
+import qualified GhcTags.ETags as ETags
+
 import           Plugin.GhcTags.Options
 import           Plugin.GhcTags.FileLock
-
-import           Plugin.GhcTags.Generate
-import           Plugin.GhcTags.Tag
-import           Plugin.GhcTags.Stream
-import qualified Plugin.GhcTags.CTags as CTags
-import qualified Plugin.GhcTags.ETags as ETags
 
 
 -- | The GhcTags plugin.  It will run for every compiled module and have access
